+++ ir/be/beutil.h
@@ -37,7 +37,7 @@ static inline const ir_node *get_block_const(const ir_node *irn)
 /**
  * Gets the Proj with number pn from irn.
  */
-ir_node *be_get_Proj_for_pn(const ir_node *irn, long pn);
+ir_node *be_get_Proj_for_pn(const ir_node *irn, unsigned pn);
 
 /**
  * Returns an array (an ARR_F) of the programs blocks in reverse postorder
